Rename SCSS files to use incident
