+++ src/pages/case/[id].tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/router"
 import Link from "next/link"
 import { BreadcrumbBar, Breadcrumb } from "@trussworks/react-uswds"
 
-import styles from "styles/case.module.scss"
+import styles from "styles/incident.module.scss"
 import { useCase } from "hooks/cases"
 
 const CasesPage: NextPage = () => {
@@ -14,7 +14,7 @@ const CasesPage: NextPage = () => {
   const { data } = useCase(id)
 
   return (
-    <main className={styles.case}>
+    <main className={styles.incident}>
       <header>
         <BreadcrumbBar className={styles.breadcrumbs}>
           <Breadcrumb>
