+++ src/styles/incident.module.scss
@@ -1,6 +1,6 @@
 @import "styles/uswds";
 
-.case {
+.incident {
   header {
     padding: 0 4rem;
     @include u-bg($theme-color-base-lightest);
