+++ src/styles/incidents.module.scss
@@ -1,6 +1,6 @@
 @import "styles/uswds";
 
-.cases {
+.incidents {
   padding: 2rem 4rem;
 
   h1 {
